branches/mbutscher/work: 2.0beta05
* Option "footnotes as wikiwords" moved to different
  option page
* In "choose wikiword" dialog (for choosing parents,
  children...), pages can be sorted alphabetically
* Option to change tree font
* More detailed instruction in error log in case of
  bad AppLock.lock file

* Windows: Show error message if copying of files
  fails during export
* Windows: New try to make crashes more seldom when
  exiting from system tray
* Windows: Generated file URLs now use ':' after
  drive letter instead of '|'

* Internal: Class AbstractDocPage between DocPage
  and DataCarryingPage to prepare for versioning
* Internal: For insertions, AST-node object gets
  member "astNodeStack" to deliver context information
* Internal: New function
  MainApp.addWikiWikiLangOptionsDlgPanel() to allow
  wiki language specific option panels




branches/stable-1.9:
1.9rc06

* Bug fixed: Error log could be cluttered up with
  messages because of bad time/date format
* More detailed instruction in error log in case of
  bad AppLock.lock file

* Windows: Show error message if copying of files
  fails during export
* Windows: New try to make crashes more seldom when
  exiting from system tray
* Windows: Generated file URLs now use ':' after
  drive letter instead of '|'

+++ lib/pwiki/wikidata/original_sqlite/WikiData.py
@@ -181,23 +181,10 @@ class WikiData:
             try:
                 filePath = self.getWikiWordFileName(word)
             except WikiFileNotFoundException:
-                if self.wikiDocument.getWikiConfig().getboolean("main",
-                        "wikiPageFiles_gracefulOutsideAddAndRemove", True) and \
-                        self._guessWikiWordFileName(word) is not None:
-                    # Page not in database but appropriate page file
-                    # present in data directory -> refresh database and try again
-
-                    self.refreshDefinedContentNames(deleteFully=True)
-                    filePath = self.getWikiWordFileName(word)
-                else:
-                    raise
-
-            if self.wikiDocument.getWikiConfig().getboolean("main",
-                    "wikiPageFiles_gracefulOutsideAddAndRemove", True) and \
-                    not os.path.exists(filePath):
-                # File is missing and this should be handled gracefully
-                self.refreshDefinedContentNames(deleteFully=True)
-                return u""
+                raise
+#                 if self.wikiDocument.getWikiConfig().getboolean("main",
+#                         "wikiPageFiles_gracefulOutsideAddAndRemove", True):
+#                     return u""
 
             content = loadEntireTxtFile(filePath)
 
@@ -240,6 +227,17 @@ class WikiData:
                 self.connWrap.execSql("update wikiwords set modified = ? "
                         "where word = ?", (moddate, word))
 
+                if self.wikiDocument.getWikiConfig().getboolean("main",
+                        "wikiPageFiles_gracefulOutsideAddAndRemove", True):
+                    try:
+                        self.getWikiWordFileName(word, mustExist=True)
+                    except WikiFileNotFoundException:
+                        fileName = self.createWikiWordFileName(word)
+
+                        self.connWrap.execSql("update wikiwords set filepath = ?, "
+                                "filenamelowercase = ? where word = ?",
+                                (fileName, fileName.lower(), word))
+
             self.cachedContentNames = None
         except (IOError, OSError, sqlite.Error), e:
             traceback.print_exc()
@@ -256,7 +254,7 @@ class WikiData:
         try:
             self._updatePageEntry(word, moddate, creadate)
 
-            filePath = self.getWikiWordFileName(word)
+            filePath = self.getWikiWordFileName(word, mustExist=False)
             writeEntireFile(filePath, content, True)
 
             fileSig = getFileSignatureBlock(filePath)
@@ -276,6 +274,9 @@ class WikiData:
         """
         try:
             oldFilePath = self.getWikiWordFileNameRaw(oldWord)
+            # To throw exception in case of error
+            self.getWikiWordFileName(oldWord)
+
             head, oldFileName = os.path.split(oldFilePath)
 #             head = pathDec(head)
 #             oldFileName = pathDec(oldFileName)
@@ -311,7 +312,7 @@ class WikiData:
             self.connWrap.execSql("delete from wikiwords where word = ?",
                     (word,))
             self.cachedContentNames = None
-            if exists(fileName):
+            if fileName is not None and os.path.exists(fileName):
                 os.unlink(fileName)
         except (IOError, OSError, sqlite.Error), e:
             traceback.print_exc()
@@ -446,10 +447,13 @@ class WikiData:
             raise DbWriteAccessError(e)
 
 
-    def deleteWord(self, word):
+    def deleteWord(self, word, delContent=True):
         """
         delete everything about the wikiword passed in. an exception is raised
         if you try and delete the wiki root node.
+        
+        delContent -- Should actual content be deleted as well? (Parameter is
+                not part of official API)
         """
         if word != self.wikiDocument.getWikiName():
             try:
@@ -462,7 +466,8 @@ class WikiData:
                     self.deleteChildRelationships(word)
                     self.deleteProperties(word)
                     self.deleteTodos(word)
-                    self.deleteContent(word)
+                    if delContent:
+                        self.deleteContent(word)
                     self.deleteWikiWordMatchTerms(word, syncUpdate=False)
                     self.deleteWikiWordMatchTerms(word, syncUpdate=True)
                     self.connWrap.commit()
@@ -533,21 +538,18 @@ class WikiData:
                 filePath = self.getWikiWordFileName(word)
             except WikiFileNotFoundException:
                 if self.wikiDocument.getWikiConfig().getboolean("main",
-                        "wikiPageFiles_gracefulOutsideAddAndRemove", True) and \
-                        self._guessWikiWordFileName(word) is not None:
-
-                    self.refreshDefinedContentNames(deleteFully=True)
-                    # Try again
-                    filePath = self.getWikiWordFileName(word)
+                        "wikiPageFiles_gracefulOutsideAddAndRemove", True):
+                    # File is missing and this should be handled gracefully
+                    return True
                 else:
                     raise
 
-            if self.wikiDocument.getWikiConfig().getboolean("main",
-                    "wikiPageFiles_gracefulOutsideAddAndRemove", True) and \
-                    not os.path.exists(filePath):
-                # File is missing and this should be handled gracefully
-                self.refreshDefinedContentNames(deleteFully=True)
-                return True
+#             if self.wikiDocument.getWikiConfig().getboolean("main",
+#                     "wikiPageFiles_gracefulOutsideAddAndRemove", True) and \
+#                     not os.path.exists(filePath):
+#                 # File is missing and this should be handled gracefully
+#                 self.refreshDefinedContentNames(deleteFully=True)
+#                 return True
 
             fileSig = getFileSignatureBlock(filePath)
 
@@ -1026,17 +1028,19 @@ class WikiData:
             # Delete words for which no file is present anymore
             for path in self.connWrap.execSqlQuerySingleColumn(
                     "select filepath from wikiwords"):
-                if not os.path.exists(longPathEnc(os.path.join(self.dataDir, path))):
-                    if not deleteFully:
-                        self.connWrap.execSql("delete from wikiwords "
-                                "where filepath = ?", (path,))
-                        self.commitNeeded = True
-                    else:
+
+                testPath = longPathEnc(os.path.join(self.dataDir, path))
+                if not os.path.exists(testPath) or not os.path.isfile(testPath):
+                    if deleteFully:
                         words = self.connWrap.execSqlQuerySingleColumn(
                                 "select word from wikiwords "
                                 "where filepath = ?", (path,))
                         for word in words:
-                            self.deleteWord(word)
+                            self.deleteWord(word, delContent=False)
+
+                    self.connWrap.execSql("delete from wikiwords "
+                            "where filepath = ?", (path,))
+                    self.commitNeeded = True
 
             # Add new words:
             ti = time()
@@ -1051,8 +1055,8 @@ class WikiData:
                     
                     self.connWrap.execSql("insert into wikiwords(word, created, "
                             "modified, filepath, filenamelowercase, "
-                            "filesignature) "
-                            "values (?, ?, ?, ?, ?, ?)",
+                            "filesignature, metadataprocessed) "
+                            "values (?, ?, ?, ?, ?, ?, 0)",
                             (wikiWord, ti, st.st_mtime, path, path.lower(),
                                     sqlite.Binary(fileSig)))
         except (IOError, OSError, sqlite.Error), e:
@@ -1125,18 +1129,44 @@ class WikiData:
 
         if path is None:
             raise WikiFileNotFoundException(
-                    _(u"Wiki page not found for word: %s") % wikiWord)
+                    _(u"Wiki page not found (no path information) for word: %s") %
+                    wikiWord)
 
         return path
 
 
-    def getWikiWordFileName(self, wikiWord):
+    def getWikiWordFileName(self, wikiWord, mustExist=True):
         """
         Not part of public API!
         Function must work for read-only wiki.
         """
-        path = self.getWikiWordFileNameRaw(wikiWord)
-        return longPathEnc(join(self.dataDir, path))
+        try:
+            path = longPathEnc(join(self.dataDir,
+                    self.getWikiWordFileNameRaw(wikiWord)))
+    
+            if mustExist and \
+                    (not os.path.exists(path) or not os.path.isfile(path)):
+                 raise WikiFileNotFoundException(
+                        _(u"Wiki page not found (bad path information) for word: %s") %
+                        wikiWord)
+        except WikiFileNotFoundException:
+            if self.wikiDocument.getWikiConfig().getboolean("main",
+                    "wikiPageFiles_gracefulOutsideAddAndRemove", True):
+                # Refresh content names and try again
+                self.refreshDefinedContentNames(deleteFully=True)
+            
+                path = longPathEnc(join(self.dataDir,
+                        self.getWikiWordFileNameRaw(wikiWord)))
+        
+                if mustExist and \
+                        (not os.path.exists(path) or not os.path.isfile(path)):
+                     raise WikiFileNotFoundException(
+                            _(u"Wiki page not found (bad path information) for word: %s") %
+                            wikiWord)
+            else:
+                raise
+
+        return path
 
 
     def createWikiWordFileName(self, wikiWord):
@@ -1774,7 +1804,7 @@ class WikiData:
             
             if filePath is None:
                 return None  # TODO exception?
-            
+
 
             datablock = loadEntireFile(join(self.dataDir, filePath))
             return datablock
